Track resources that are pending external name assignment
Some Crossplane resources have non-deterministic external
names that are returned at Create time. We must record those
names and rely on them to determine whether the external
resources exist during subsequent Observe calls.

The absence of an external name does not guarantee that we
didn't create an external resource. It's possible that we
created the resource but failed to update its name. It's
also possible that we did update the name but have since
read a stale, unnamed, version of the resource from cache.

To deal with this we set a 'pending external name' annotation
immediately before creating our external resource. If we get
to the Create call because we have a stale view of our managed
resource the Update call that persists the annotation will fail.
If we are already pending external name assignment when we get
to the Create call we know it's possible that we created an
external resource but did not persist its name, so we refuse to
proceed rather than creating a duplicate.

Signed-off-by: Nic Cope <negz@rk0n.org>